+++ realtime/utils.js
@@ -1,11 +1,3 @@
-function get_hostname(url) {
-	if (!url) return undefined;
-	if (url.indexOf("://") > -1) {
-		url = url.split("/")[2];
-	}
-	return url.match(/:/g) ? url.slice(0, url.indexOf(":")) : url;
-}
-
 function get_url(socket, path) {
 	if (!path) {
 		path = "";
@@ -15,5 +7,4 @@ function get_url(socket, path) {
 
 module.exports = {
 	get_url,
-	get_hostname,
 };
